Fix test which assumes `toString` of `NullThrownError()`.
The `NullThrownError` is deprecated (no longer possible in
null safe Dart, and is already being replaced by a `TypeError`.)

A test assumes that the `toString` of a `NullThrownError` is
`Instance of 'NullThrownError'`, where it's now
`Instance of 'TypeError'`.
In general, we do not promise that `toString` of error objects
are stable.