Bump min dependency on code_builder
Towards dart-lang/mockito#561

This package depends on a new API that was introduced in version 4.2.0 of
`code_builder`, but we did not document the API in the changelog or bump min
dependencies.

Changing this now won't retroactively correct the fact that users with an older
code_builder can pick up an older mockito, but it should reduce the risk of
users continuing to see it.

PiperOrigin-RevId: 465185748