Avoid `Future.value(null)` as a dummy return value.
We allow stubbing a method with a potentially non-nullable type argument
argument because we can use a dummy value of `Future.value(null)`. This object
is never awaited by mockito, and never exposed to the user, so it is
more-or-less safe. However, `Future<T>.value(null)` is now statically invalid
for a potentially non-nullable `T`, so it is cleaner if we avoid it in mockito.

Subtyping `Future` is allowed, so let's just do it.

PiperOrigin-RevId: 483448979