Merge test selections by path
Closes #1785

When passing the same path multiple times on the command line, with
different `name`, `line`, or `col`, options run the test suite only
once. This does not affect cases where the same suite is loaded through
multiple path arguments from the command line, for instance if it was
specified directly and also loaded in a directory.

Remove the single `patterns`, `line`, and `col`, fields from
`SuiteConfiguration` in favor of a set of a new `TestSelection` with
these fields. Within each selection _all_ criteria must match for a test
to run, but test selections can independently allow a test to run.

Drop the `PathConfiguration` class, in favor of a map from the path to a
set of test selections. The global `Configuration` gets the map, and
eventually each suite run gets it's applicable set of selections.
Replace `Configuration.paths` with a map from the path option to a set
of selections. Group test path configurations by path to collect test
selections when parsing command line arguments.

Don't allow the test selections for a `SuiteConfiguration` in any of
the constructors - only accept a value with a new `selectTests` method
so that it is more clear where the data flows.

Also, add a `globalPatterns` field on `Configuration` to work alongside
test selections, but maintain separation between data sources.
Previously the patterns would be copied from the `suiteDefaults` for
each suite, and the suite specific patterns would be merged in.

Check whether any selection uses line or column when deciding whether to
capture test definition traces.

Add a test in `line_and_col_test.dart` that using the same test path and
line multiple times results in a single test ran.