Use double quotes for test names on windows (#1802)
Closes #1798

The windows shell does not allow single quotes around arguments. Use
double quotes on windows, and escape double quotes with 3 double quotes
in a row. The escape sequence closes the quoted string, emits an
adjacent double quote, then opens a new quoted string.

If any quotes are escaped the command will not work for users of the
bash shell in VSCode, but there were already issue with the path
separators in that shell so this shouldn't introduce severe new
problems. Test names that don't have double quotes in them should work
in either place.