Fix missing label and reason after isNotNull (#1797)
When creating a context which is an alias of another, at the same level
in the context tree, the parent was copied. When this happened on the
root in the tree, the alias would also appear to be a root, but it would
be missing some other important information from the root, like the root
label and reason. Separately, the reason which should have always been
read from the root, was not.

- Rename the first constructor from `_` to `_root` to more clearly
  indicate where should be used.
- Add a doc that only the root constructor should have no parent.
- Treat the aliased context as the parent, instead of copying the
  originals parent. This only matters when the aliased context is the
  root, but it does not hurt for the parent references to have an extra
  hop in other cases.
- Read the reason from the root instead of on the check that failed.
- Add some initial tests for the message in a thrown `TestFailure`. Add
  tests for the `Expected:` label and `Reason:` being retained when
  `isNotNull()` is used on the root context, which means nesting
  `atSameLevel` and creating an aliased root context.