Stop using deprecated errors (dart-lang/matcher#201)
Prepare for the errors to be removed from the SDK and stop referencing
them entirely. Deprecate the matchers for these errors and change the
types to target the more general `Error` and `TypeError` which will
still catch the same situations.

Move the ignore for same package deprecations to cover the entire test
file.