Add FailureDetail to allow expressive failures
Rejections do not carry any context about where in the
Check tree a failure occurred, so they are no sufficient to describe the
difference between the expected an actual values like we use in the
messages for test failure exceptions.

Some conditions which do a `softCheck` may want to print more
detailed mismatch information, for example an `Iterable.every` check may
want to print why an element didn't match, in addition to which index
failed.

Change the `fail` callback from taking a formatted String message and
the Rejection, to taking a `CheckFailure` with the rejection and a lazy
details field. The `FailureDetail` has the `expected` and `actual` that
were previously computed in separate methods, as well as information
about how far to indent lines from the `Rejection` so they line up with
the label for the specific check that failed.

Change the `softCheck` methods to return `CheckFailure?` instead of
`Rejection?`. Update conditions which were run against these values to
target `Check<CheckFailure?>` and add an extra `has` to read the
`rejection` field.

Remove the `_reason` field from `_TestContext` and handle it entirely
within `checkThat` since it now constructs the failure message.

Add an optional argument to the `indent` utility to let it indent more
than one level at a time.