Add pairwiseCompare condition
Re-implements `pairwiseCompare` from `matcher` which has limited, but
non-zero usage.

The callback gets a `Check` and the expected value, instead of the
actual and expected values, and causes a condition to fail instead of
returning `false` for mismatch.

Rejects iterables that have the wrong number of elements, or that fail
an individual condition for any element.