Implement Future.doesNotComplete
Add `Context.lateReject` to allow a condition to report a failure after
it has already returned `null` (instead of a `Rejection`). This will
also be useful for a replacement for `expectAsync` with a limit on
allowed calls.

Add `doesNotComplete` in the Future extension. Always accept the value
synchronously, but reject it late if the Future does complete. Include
the stack trace for futures that result in an error.