Tighten types of Extracted constructors (#1829)
The intention is that one of the fields will always be non-null (or at
least if `T` isn't nullable) so each constructor takes one of the fields
as an argument and sets the other to null. Tighten the argument types
from the nullable field types to actually enforce this. Add a local
variable to allow promotion in one location.