Add Future.doesNotComplete condition (#1827)
Add `Context.expectUnawaited` to allow a condition to report a failure
through a callback instead of having to complete a Future as success or
failure. This will also be useful for a replacement for `expectAsync`
with a limit on allowed calls.

Add `doesNotComplete` in the Future extension. Always accept the value
synchronously, but reject it asynchronously if the Future does complete.
Include the stack trace for futures that result in an error.