Add Condition abstraction
Previously, all "condition" arguments were callbacks that take a
`Check`. Now there is a `Condition` class which has an `apply` method
for the same purpose.

Add an `it()` Utility to get an object that is both a `Condition`, and a
`Check`. Expectations run against the check are recorded and played back
against the real check when the condition is applied. Expectations can
be chained with a cascade, so that the outermost object is the one
passed as an argument.

Add a factory constructor which takes a callback because there is
currently no way to express a numeric comparison expectation with a
cascade. This wouldn't be necessary if we had methods like
`.isGreaterThan` instead of `operator >`.

Move `skip` to an extension since there are now multiple implementations
of `Check`.