Improve output for collections and long strings
Change the `actual` argument for rejections from a single String to an
`Iterable<String>`. This allows for maintaining alignment when a multi
line value is displayed with indentation. Change the `literal` utility
to return an Iterable, and update conditions passing something other
than the return from `literal` to pass a List to `Rejection.actual`.

Add a `postfixLast` utility to mirror `prefixFirst` and append content
to the last line of a value. Combining `prefixFirst` and `postfixLast`
with delimiters like quotes or braces allows wrapping content than maybe
be 1 or more lines.

Flesh out the `literal` utility. The implementation is based on
`prettyPrint` from `matcher` with some simplifications and adapted to
return an Iterable instead of a String with newlines. One significant
difference from `prettyPrint` is that type names are not included since
they are often redundant.

In one place where `literal` was used for a signature that must take a
single String (the label in a nested context can only be a single line)
join the result with `r'\n'`.