Avoid empty `::group::` in Github reporter
A `::group::` will have an arrow in the Github UI. This is misleading in
the majority of cases where there are no extra lines, and masks the
cases where there is output from the test.

Only wrap in `::group::` when there is a non-empty set of messages or
errors, since those indicate that there will be additional lines in the
output.