Prevent infinite loops for recursive collections
Add a max depth of 1000. Throw an exception if a search would ever
progress deeper than the max depth from the top of the collection. Catch
the exception and report it as a failure. Typically a failure shouldn't
be triggered by an exception within a condition extension - but since
this should be a rare failure which indicates an error with the test, an
exception is an OK shortcut.