Add unordered iterable comparisons
Pull the common parts of `_findSetDifference` and `_findMapDifference`
into a utility called `unorderedCompared` which can be reused for
`unorderedEquals` and `unorderedMatches`. This removes some specificity
from rejections for non-matching Maps - it will now always refer to an
unmatched entry in its entirety, and will not have a special message for
when there is a key with no matches, vs when there are matching keys but
those keys have non-matching values.

Change `_hasPerfectMatching` to `_findUnpaired`. Don't require that the
actual and expected vertex sets have the same cardinality, take the
cardinality of the actual vertex set as an argument. Return list of the
unmatched vertices in both sets. Allows more specific failures than
solely reporting unequal lengths or that the elements don't have a full
pairing - we can be specific about at least one missing or unexpected
element.

Add `unorderedEquals` and `unorderedMatches` conditions for Iterable.