Overwrite stack traces to expectation call stack
When applying a `Condition` (as opposed to using `softCheck`) the stack
trace includes some uninteresting checks infrastructure, and the line
from the `_test.dart` file typically points to the higher level
expectation which took the Condition as an argument. It is often more
useful to point to the specific expectation called on the Condition
which is failing.

Record `StackTrace.current` for all direct interactions with
`_ReplayContext`. When replaying the interaction on another context,
catch exceptions and rewrite them to the recorded stack trace.

Add an expando to remember which exceptions have already had their
stack traces rewritten. The callback for a check which causes nesting
also runs the interactions for the nested context, but we don't want the
stack to point to the nesting level. Only the inner nested replay
context should rewrite the stack trace.

When replaying onto a replay context (`it()..that(someCondition)`),
directly copy the interactions and original stack traces instead of
calling the method and getting new stack traces. This makes it possible
to use condition syntax inside expectation definitions without
overwriting the stack traces to the utility from the test file, and
should be more efficient.