Overhaul async matchers
Start a transaction before checking expectations. If the check fails,
reject the transaction. Add tests for the transaction behavior of each
condition.

Add `emitsError` to check that the queue has an error of a specified
type instead of an event.

Add `emitsInOrder`. This API mainly solves the problem of checking
multiple async expectations against the same `Check` instance, since it
is no possible to `await` in between cascades. It also maintains a
transaction for the entirety of the conditions, so that the stream gets
rolled all the way back on failure. The conditions use will use async
expectations, so we cannot synchronously find a description for all the
conditions. Descriptions are collected while checking the stream when we
have room for async work. The `clause` callback is not called until
after the `predicate` callback has completed. During `describeAsync` the
`predicate` callback is never called, so for the describe use case the
clause will fall back to stating the number of conditions.

Add `emitsAnyOf` to check multiple possible stream scenarios from the
same point in the queue. Similar to `emitsInOrder` it uses a fallback
for describe, and collects condition descriptions when the predicate
callback is run. On rejection describes the rejection for each
individual condition choice.

Add `mayEmit` and `mayEmitMultiple` to consume events from the queue
without ever failing.

Add `isDone` to check that the queue ends without more events or
errors.

Expose `softCheckAsync` and `describeAsync` for defining expectation
extensions. Previously it was imported from src only for tests.

Fix up some users facing strings so that they read nicely in more
situations. Add a generic bound for `throws` and change the label based
on whether a type was passed. When writing the label for nesting clauses
without sub-clauses, omit the `that:` postfix.

Add `hasDescription` and `hasDescriptionAsync` test utilities. Add tests
for the descriptions of the async conditions.