Refactor tests to a new isRejectedBy utility (#1867)
The new structure has a few advantages over the previous pattern of
calling `softCheck` and passing the result to `checkThat`, we can omit
some generic type arguments, and the failure can carry more context when
the failure is caused by accepted a value that should be rejected.

With `softCheck` the generic must be explicit on the `it()` call, or
explicit on `softCheck` and then filled through downward inference.
Without an explicit generic the correct extension methods are not
available for the condition. With a value passed to `checkThat` instead,
the generic can be inferred, and then carried forward to `it()` through
downwards inference.

With `softCheck` the value and the `Condition` are both opaque to the
failure, because the check only sees the `Rejection?`. With
`isRejectedBy` we can include a description of the condition for
failures where the value was incorrectly accepted. If the failure is
instead with the `actual` or `which` conditions, this context is still
missing from the failure message.

Change `deepCollectionEquals` to return an `Iterable<String>` instead of
a `Rejection?`. The `which` is the only interesting information, and
this lets us test the method with standard expectations instead of
still needing a `isARejection`.

Strengthen the tests for the "actual" rejection arguments. Test that the
default was actually used instead of skipping the expectation entirely
for the cases where we don't pass an expectation for actual. Fill in
some non-default actual values in tests that were previously not
checking them.