Try a postfix phrasing
Make the utility to get a `Subject` an extension method. Since it reads
after the value, change the phrasing to "must" and expectations to
"beSomeDescription".

This causes chained conditions to read poorly

``dart
actualIterable.must.haveLength.beGreaterThan(0);
```

Also needs a solution for passing a "reason" at the top level, but
hopefully which prevents passing a reason anywhere else.