Experiment with disallowing subject chaining
Remove the return value from `nest` and `nestAsync`, for passing the
nested condition ahead of time.

This interacts really poorly with `has`. We should only consider doing
this if we can find a workaround to allow inference for `has` or if we
expect that the vast majority of users will use codegen and we can hide
it from them.

I do like how this eliminates the `which` calls for some async
expectations.