Rename checkThat to check (#1880)
Save 4 characters of typing millions of types. The line will not read as
nicely in many situations, but it doesn't make a significant difference.

The shorter name, with no capital letters, also draws the least possible
attention to this name, which carries the least information of any part
of the expression.