Rename some extensions for consistency
These are technically user facing, but not likely to be referenced.

Rename `ThrowsCheck` to `FunctionChecks`, and `NullabilityChecks` to
`NullableChecks`. All extensions that define expectation methods are
named `*Checks` based on the type they support, plus `CoreChecks` for
any type.

Rename `Skip` to `SkipExtension` to avoid conflict with `Skip` from
`package:test`.

Rename `StreamQueueWrap` to `WithQueueExtension`.