Support running tests by absolute file: uri (#1893)
Fixes https://github.com/dart-lang/test/issues/1891
Fixes https://github.com/dart-lang/test/issues/1899

For any path which looks like it has a scheme (contains "://"), parse it as a uri and grab the path from that. On windows, we also remove any leading `/` from the path because it will be followed by a drive letter, which we want to be the first part of the path (and not preceded by a `/`).