Rename matches to matchesPattern, accept Pattern (#1907)
This will treat a `String` argument differently from the equivalent
matcher in `package:matcher`, which would have converted strings to
regex. Document this in the migration guide.

Add tests showing the behavior of passing strings, and explicitly choose
cases that show the argument is not treated as a regular expression.

Expand the doc comment and show examples of both types of arguments.