Add more specific issue templates
Add a separate template for reporting a bug or requesting a feature with
the test runner.
Make the language about `package:checks` specific that the package is in
preview and we are taking any type of feedback.
Add an "other" bug for issues that may get filed by the team about
internal stuff.