Handle absolute paths on Windows (#1923)
Since the latest `package:test` rolled into the SDK, some DAP tests started failing to run tests with `dart test`. They use absolute paths on Windows: https://github.com/dart-lang/sdk/issues/51348

This updates the code to check if it looks like we are getting a path that starts with drive letter, and in that case prepend it with file:/// before parsing. I also made the logic for stripping out a leading / more robust (only do it if it really looks like a drive letter path).

@jakemac53