Add async examples to README (#1946)
Towards #1928

Add a section in the README about async expectations. Give examples of
awaiting the result, and describe that when the result is not awaited
the test runner is still aware of the work. Describe that some
expectations cannot be awaited.

Describe the reason that `StreamQueue` is used for stream expectations
and give examples of using the `withQueue` and a manual wrapping.

Mention that both `expect` and `expectLater` are migrated to the same
`check` call in the migration guide. Give an example of migrated an
awaited expectation.