Drop the Condition abstraction
Explore what tests would look like if we remove this concept entirely.

Tests are migrated to use `it` as the subject name everywhere for
convenience, but tests in the real work may be more likely to choose a
single letter or longer name based on the value, for instance `e` for
the `element` checked in an iterable.

Some linting setups may be able to statically help avoid passing a
`Future<void> Function(Subject)` where a synchronous condition is
required. This static checking couldn't work for the `it()` syntax.

When inference fails, for instance when passing a callback in a
collection for `deepEquals`, there is no static help, and at runtime
there will be errors about `Subject` not having instance methods for the
extensions that are used.

Add special handling in `literal` for `void Function(Subject)`
instances.