Fix timeouts when a predicate throws instead
A predicate or extraction callback is not supposed to throw - it should
return a Rejection to indicate a failure - however if an exception is
thrown the test should fail for that exception and immediately be done.

Move the `outstandingWork.complete()` call to a `finally` block so it
is never skipped. Move it after the nested condition is applied to also
fix a potential bug if the `Condition` is a callback that does any async
work outside calls to the context. This isn't a typical pattern, but
could become more typical if we drop the `Condition` abstraction.