More smoothly handle missing compiler info (#1980)
The constructor signature uses a nullable argument to smooth over the
compatibility across versions which allowed us to land in monolithic
repositories without needing to roll `test_api` and `flutter_test`
atomically. This extends the same backwards compatibility to the json
API used between the runner and tests to smooth over compatibility with
`flutter test` which may try to run test compiled against a newer API
than it uses to communicate with them.