Remove deprecation of test_api top level libraries (#1994)
Closes #1990

The `scaffolding` library should not be deprecated to avoid marking the
useful APIs as deprecated during autocomplete.

The `backend` library is no longer deprecated because it _is_ intended
for use in `flutter_test`, and the deprecation isn't likely to be
blocking any uses.

Export `registerException` from the scaffolding library. This had
originally been held back from the focused libraries because it isn't a
great API, but it's too entrenched to be worth deprecating at this
point, so it may as well be fully supported.