Migrate to Dart 3
Bump the min SDK to `3.0.0` across the test packages. It will not be
possible to test or run on older SDKs once we migrate to the next
version of `analyzer`.

Drop the `legacy_tests`. We will continue to have coverage for that
behavior with internal tests.

Add class modifiers to classes exposed through the public APIs of
`test`, `test_core`, or `test_api`. Internal classes that are not
exported through a public library have not been considered, and more
modifiers may be useful for these classes in the future.

Deprecate the export of `Result` and `Status` from
`test_api/hooks_testing.dart`. These had been part of the API design
before the `State` enum was added, but the export was mistakenly left
in.

Deprecate the export of `RemoteListener` from a `lib/src` import which
was maintained for compatibility for existing imports of that URI. The
next breaking change for `test_api` should remove the library.