Migrate to Dart 3 (#2024)
Bump the min SDK to `3.0.0` across the test packages. It will not be
possible to test or run on older SDKs once we migrate to the next
version of `analyzer`.

Drop the `legacy_tests`. We will continue to have coverage for that
behavior with internal tests.

Add class modifiers to classes exposed through the public APIs of
`test`, `test_core`, or `test_api`. Internal classes that are not
exported through a public library have not been considered, and more
modifiers may be useful for these classes in the future.

Add class modifiers to the classes exposed from `checks`.

Deprecate the export of `Result` and `Status` from
`test_api/hooks_testing.dart`. These had been part of the API design
before the `State` enum was added, but the export was mistakenly left
in.

Change the `Compiler` class into an `enum`.

Make `Fake` a `mixin class` to avoid breaking existing usage.