Fix root suite comparison on Windows (#2038)
Fixes #2037

Canonicalize paths when checking whether a stack frame is coming from
the test suite root.

Add a test case which runs with an absolute path argument. On windows,
this situation would have resulted in the `root_` location information
being missing because of a difference in the stack frame paths and the
suite path before canonicalization. This test fails on windows without
the canonicalization.

Make the test utils more flexible to handle the new output differences
across operation systems.
Allow a `Matcher` value for the `path` argument to `suiteJson`. This
optional argument had previously been unused. Use `equalsIgnoreCase`
because on windows there is a difference in the case of the drive
letter.
Allow a `Matcher` value for the `name` argument to `testStartJson`. Take
an `Object` to allow either `Matcher` or `String` for existing uses.
Check the "loading" output line loosely by checking only the start and
end.