Remove "compiling" messages
All platforms will use the "loading" message. This had been a heuristic
for the dart2js tests because originally it was the standout amongst the
otherwise run-form-source options with the VM and dartium.

The number of platforms has since expanded, and flutter appears to use
as a dart2js platform that isn't precompiled, because we have a
different way to trigger the precompiled path than `flutter test` which
precompiles always.