More loosely handle error message from SDK (#2064)
A recent dev version of the SDK changed the formatting for a regex parse
error that we were depending on in a test - a space was introduced where
it was missing between "group" and the text of the group, "Unterminated
group (foo". The test is using the `contains` matcher for the values in
`containsInOrder`, so we can make the expectation looser to allow either
format.