Remove "compiling" messages (#2063)
All platforms will use the "loading" message. This had been a heuristic
for the dart2js tests because originally it was the standout amongst the
otherwise run-from-source options with the VM and dartium.

The number of platforms has since expanded, and flutter appears as a
dart2js platform that isn't precompiled, because the test runner has a
different way to trigger the precompiled path than `flutter test` which
precompiles always.

This will have a user facing impact - users will see "loading" instead of
"compiling" and the message will show up more often, including when
there are no dart2js tests running.